+++ sdk/cosmos/azure_data_cosmos/src/clients/database_client.rs
@@ -1,8 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-use crate::cosmos_request::CosmosRequest;
-use crate::operation_context::OperationType;
 use crate::routing::global_endpoint_manager::GlobalEndpointManager;
 use crate::routing::global_partition_endpoint_manager::GlobalPartitionEndpointManager;
 use crate::{
@@ -15,6 +13,7 @@ use crate::{
     ThroughputOptions,
 };
 use azure_core::http::Context;
+use azure_data_cosmos_driver::models::{CosmosOperation, DatabaseReference};
 use azure_data_cosmos_driver::CosmosDriver;
 use std::sync::Arc;
 
@@ -29,6 +28,7 @@ pub struct DatabaseClient {
     database_id: String,
     pipeline: Arc<GatewayPipeline>,
     driver: Arc<CosmosDriver>,
+    database_ref: DatabaseReference,
     global_endpoint_manager: Arc<GlobalEndpointManager>,
     global_partition_endpoint_manager: Arc<GlobalPartitionEndpointManager>,
 }
@@ -44,6 +44,8 @@ impl DatabaseClient {
         let database_id = database_id.to_string();
         let link = ResourceLink::root(ResourceType::Databases).item(&database_id);
         let containers_link = link.feed(ResourceType::Containers);
+        let database_ref =
+            DatabaseReference::from_name(driver.account().clone(), database_id.clone());
 
         Self {
             link,
@@ -51,6 +53,7 @@ impl DatabaseClient {
             database_id,
             pipeline,
             driver,
+            database_ref,
             global_endpoint_manager,
             global_partition_endpoint_manager,
         }
@@ -103,17 +106,21 @@ impl DatabaseClient {
     ///     .into_model()?;
     /// # }
     /// ```
-    #[allow(unused_variables, reason = "This parameter may be used in the future")]
     pub async fn read(
         &self,
         options: Option<ReadDatabaseOptions>,
     ) -> azure_core::Result<ResourceResponse<DatabaseProperties>> {
-        let cosmos_request = CosmosRequest::builder(OperationType::Read, self.link.clone()).build();
+        let options = options.unwrap_or_default();
+        let operation = CosmosOperation::read_database(self.database_ref.clone());
 
-        self.pipeline
-            .send(cosmos_request?, Context::default())
-            .await
-            .map(ResourceResponse::new)
+        let driver_response = self
+            .driver
+            .execute_operation(operation, options.operation)
+            .await?;
+
+        Ok(ResourceResponse::new(
+            crate::driver_bridge::driver_response_to_cosmos_response(driver_response),
+        ))
     }
 
     /// Executes a query against containers in the database.
@@ -162,23 +169,27 @@ impl DatabaseClient {
     /// # Arguments
     /// * `properties` - A [`ContainerProperties`] describing the new container.
     /// * `options` - Optional parameters for the request.
-    #[allow(unused_variables, reason = "This parameter may be used in the future")]
     pub async fn create_container(
         &self,
         properties: ContainerProperties,
         options: Option<CreateContainerOptions>,
     ) -> azure_core::Result<ResourceResponse<ContainerProperties>> {
         let options = options.unwrap_or_default();
-        let cosmos_request =
-            CosmosRequest::builder(OperationType::Create, self.containers_link.clone())
-                .request_headers(&options.throughput)
-                .json(&properties)
-                .build()?;
+        let body = serde_json::to_vec(&properties)?;
+        let operation =
+            CosmosOperation::create_container(self.database_ref.clone()).with_body(body);
 
-        self.pipeline
-            .send(cosmos_request, Context::default())
-            .await
-            .map(ResourceResponse::new)
+        let operation_options =
+            crate::driver_bridge::apply_throughput_headers(options.operation, &options.throughput)?;
+
+        let driver_response = self
+            .driver
+            .execute_operation(operation, operation_options)
+            .await?;
+
+        Ok(ResourceResponse::new(
+            crate::driver_bridge::driver_response_to_cosmos_response(driver_response),
+        ))
     }
 
     /// Deletes this database.
@@ -187,17 +198,21 @@ impl DatabaseClient {
     ///
     /// # Arguments
     /// * `options` - Optional parameters for the request.
-    #[allow(unused_variables, reason = "This parameter may be used in the future")]
     pub async fn delete(
         &self,
         options: Option<DeleteDatabaseOptions>,
     ) -> azure_core::Result<ResourceResponse<()>> {
-        let cosmos_request =
-            CosmosRequest::builder(OperationType::Delete, self.link.clone()).build();
-        self.pipeline
-            .send(cosmos_request?, Context::default())
-            .await
-            .map(ResourceResponse::new)
+        let options = options.unwrap_or_default();
+        let operation = CosmosOperation::delete_database(self.database_ref.clone());
+
+        let driver_response = self
+            .driver
+            .execute_operation(operation, options.operation)
+            .await?;
+
+        Ok(ResourceResponse::new(
+            crate::driver_bridge::driver_response_to_cosmos_response(driver_response),
+        ))
     }
 
     /// Reads database throughput properties, if any.